+++ src/Dropdown.php
@@ -2067,40 +2067,6 @@ JAVASCRIPT;
         ]);
     }
 
-
-    /**
-     * Toggle view in LDAP user import/synchro between no restriction and date restriction
-     *
-     * @param $enabled (default 0)
-     **/
-    public static function showAdvanceDateRestrictionSwitch($enabled = 0)
-    {
-        /** @var array $CFG_GLPI */
-        global $CFG_GLPI;
-
-        $rand = mt_rand();
-        $url  = $CFG_GLPI["root_doc"] . "/ajax/ldapdaterestriction.php";
-        echo "<script type='text/javascript' >";
-        echo "function activateRestriction() {";
-         $params = ['enabled' => 1];
-         Ajax::updateItemJsCode('date_restriction', $url, $params);
-        echo "};";
-
-        echo "function deactivateRestriction() {";
-         $params = ['enabled' => 0];
-         Ajax::updateItemJsCode('date_restriction', $url, $params);
-        echo "};";
-        echo "</script>";
-
-        echo "</table>";
-        echo "<span id='date_restriction'>";
-        $_POST['enabled'] = $enabled;
-        include(GLPI_ROOT . "/ajax/ldapdaterestriction.php");
-        echo "</span>";
-        return $rand;
-    }
-
-
     /**
      * Dropdown of values in an array
      *